Transfer to OpenAIGym Framework
Changes to be committed:
	modified:   Grid_Soccer.py
	modified:   main.py

Modifies the existing code to fit within the OpenAI gym framework.
Specifically, it modifies Grid_Soccer.py to create a third class,
class GridBallWorld, which contains the previous agent and gridworld
classes. The new class contains the Gym api methods
- reset()
    Return None
- step(action)
    Return state, reward, done, info
Actions are passed into the step function, applied in world, and the
results are returned to the interacting learning algorithm.

Tests in main.py have been updated, through testing has yet to be done.